+++ paging-samples/src/commonMain/kotlin/ua/wwind/paging/sample/domain/repository/UserRepository.kt
@@ -3,11 +3,11 @@ package ua.wwind.paging.sample.domain.repository
 import ua.wwind.paging.sample.domain.model.User
 
 /**
- * Repository interface for accessing user data
+ * Remote Data Source contract for fetching users over network.
  */
-interface UserRepository {
+interface UserRemoteDataSource {
     /**
-     * Get users with pagination support
+     * Get users with pagination support from a remote source.
      * @param offset Starting position (0-based)
      * @param limit Number of users to fetch
      * @return List of users and total count