+++ paging-samples/src/commonMain/kotlin/ua/wwind/paging/sample/presentation/components/LoadingItem.kt
@@ -19,6 +19,7 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
@@ -32,6 +33,7 @@ import androidx.compose.ui.unit.dp
  */
 @Composable
 fun LoadingItem(
+    position: Int,
     modifier: Modifier = Modifier,
 ) {
     val infiniteTransition = rememberInfiniteTransition()
@@ -101,6 +103,7 @@ fun LoadingItem(
                         .graphicsLayer(alpha = alpha)
                 )
             }
+            Text(position.toString())
         }
     }
 }