+++ paging-samples/src/commonMain/kotlin/ua/wwind/paging/sample/presentation/components/UserCard.kt
@@ -84,6 +84,7 @@ fun UserCard(
                     color = MaterialTheme.colorScheme.onSurfaceVariant
                 )
             }
+            Text(user.id.toString())
         }
     }
 }